fix compile warnings on android and other systems, and sanity fix. from libretro guys

+++ src/cpu/m68k/m68kcpu.c
@@ -32,6 +32,8 @@ static const char copyright_notice[] =
 /* ================================ INCLUDES ============================== */
 /* ======================================================================== */
 
+#include <stdint.h>
+
 extern void m68040_fpu_op0(void);
 extern void m68040_fpu_op1(void);
 
@@ -1053,7 +1055,7 @@ unsigned int m68k_context_size()
 /* Used to calculate the context size minus the system-specific pointers, for savestates */
 unsigned int m68k_context_size_no_pointers()
 {
-	return (int)&m68ki_cpu.pointer_block_divider - (int)&m68ki_cpu;
+	return (intptr_t)&m68ki_cpu.pointer_block_divider - (intptr_t)&m68ki_cpu;
 }
 
 unsigned int m68k_get_context(void* dst)
